requirements.txt
imageio==2.3.0
matplotlib
numpy
pandas
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.1
scikit-image
scikit-learn
scipy
torch
torchvision
tqdm==4.24.0
opencv-python